tests/wpt/web-platform-tests/websockets/interfaces/WebSocket/events/008.html

<!doctype html>
<title>WebSockets: listening for events with onerror</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.js?pipe=sub></script>
<meta name="variant" content="">
<meta name="variant" content="?wss">
<div id=log></div>
<script>
async_test(function(t) {
  var ws = new WebSocket(SCHEME_DOMAIN_PORT+'/');
  var run = false;
  var foo = t.step_func(function (e) {
    run = true;
    assert_equals(e.detail, 5)
  });
  ws.onerror = foo;
  var ev = document.createEvent('UIEvents');
  ev.initUIEvent('error', false, false, window, 5);
  ws.dispatchEvent(ev);
  assert_true(run);
  t.done();
});
</script>
